pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lsvit-action"
version = "0.1.0"
description = "LSViT-based video action recognition refactored from a Jupyter Notebook."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Your Name" }
]
dependencies = [
  "torch>=2.1.0",
  "torchvision>=0.16.0",
  "timm>=1.0.0",
  "numpy>=1.24.0",
  "Pillow>=10.0.0",
  "matplotlib>=3.7.0",
  "tqdm>=4.66.0",
  "gdown>=5.0.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]